# backend/api/api.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Book Lib public API
security:
  - bearerAuth: [ ]
paths:
  /api/book/add:
    post:
      tags:
        - book
      operationId: AddBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBookRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddBookResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/permissionDenied'
  /api/book/{bookID}/delete:
    post:
      tags:
        - book
      operationId: DeleteBook
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/permissionDenied'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/{bookID}/info:
    get:
      tags:
        - book
      operationId: GetBookInfo
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBookInfoResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
  /api/author/{authorID}/info:
    get:
      tags:
        - author
      operationId: GetAuthorInfo
      parameters:
        - in: path
          name: authorID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAuthorInfoResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/author/{authorID}/list:
    get:
      tags:
        - book
      operationId: ListAuthorBooks
      parameters:
        - in: path
          name: authorID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuthorBooksResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/permissionDenied'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/{bookID}/library/add:
    post:
      tags:
        - book
      operationId: AddBookToLibrary
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/{bookID}/library/contains:
    get:
      tags:
        - book
      operationId: CheckBookInLibrary
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckBookInLibraryResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/{bookID}/library/remove:
    post:
      tags:
        - book
      operationId: RemoveBookFromLibrary
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/search:
    post:
      tags:
        - book
      operationId: SearchBooks
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchBooksRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchBooksResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
  /api/author/search:
    post:
      tags:
        - author
      operationId: SearchAuthors
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAuthorsRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAuthorsResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
  /api/article/search:
    post:
      tags:
        - article
      operationId: SearchArticles
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchArticlesRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchArticlesResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
  /api/book/category/{categoryID}/list:
    get:
      tags:
        - book
      operationId: ListBooksByCategory
      parameters:
        - in: path
          name: categoryID
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBooksByCategoryResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
  /api/library/status/list:
    post:
      tags:
        - book
      operationId: ListLibraryBooksByStatus
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListLibraryBooksByStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLibraryBooksByStatusResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/permissionDenied'
  /api/book/{bookID}/status:
    post:
      tags:
        - book
      operationId: ChangeReadingStatus
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeReadingStatusRequest'
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/permissionDenied'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/{bookID}/feedback:
    post:
      tags:
        - book
      operationId: SaveBookFeedback
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveBookFeedbackRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveBookFeedbackResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/permissionDenied'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/{bookID}/feedback/list:
    get:
      tags:
        - book
      operationId: ListBookFeedback
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBookFeedbackResponseData'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
  /api/book/{bookID}/edit:
    post:
      tags:
        - book
      operationId: EditBook
      parameters:
        - in: path
          name: bookID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditBookRequest'
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/permissionDenied'
        '404':
          $ref: '#/components/responses/notFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    notFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/notFoundResponseData'
    permissionDenied:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/permissionDeniedResponseData'
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/unauthorizedResponseData'
    badRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/badRequestResponseData'
  schemas:
...
    AddBookRequest:
      type: object
      properties:
        filePath:
          type: string
        title:
          type: string
        description:
          type: string
        authorIDs:
          type: array
          items:
            type: integer
        categoryIDs:
          type: array
          items:
            type: integer
        coverPath:
          type: string
      required:
        - filePath
        - title
        - authorIDs
        - categoryIDs
    AddBookResponseData:
      type: object
      properties:
        id:
          type: integer
      required:
        - id
    GetBookInfoResponseData:
      type: object
      properties:
        book:
          $ref: '#/components/schemas/bookInfo'
      required:
        - book
...
    CheckBookInLibraryResponseData:
      type: object
      properties:
        contains:
          type: boolean
      required:
        - contains
...